Fix installation directory for plugins.
Don't rewrite the default libdir variable for installing plugins, create
a new plugindir variable and make use of that.

This is important because libraries might have to be installed in a
different directory (for instance on 64-bit multilib systems). In such a
system, it's usual to install 64-bit libraries in $prefix/lib64 rather
than $prefix/lib.